requirements.txt
# Core deep learning framework
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Medical imaging libraries
nibabel>=3.2.0
SimpleITK>=2.1.0
pydicom>=2.3.0
scikit-image>=0.19.0

# Data processing and analysis
numpy>=1.21.0
pandas>=1.4.0
scipy>=1.8.0
scikit-learn>=1.1.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
ipywidgets>=7.6.0

# Web framework for interface
flask>=2.2.0
flask-cors>=3.0.10

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0
tensorboard>=2.9.0
wandb>=0.13.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.15.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.6.0

# Additional utilities
albumentations>=1.2.0
monai>=1.0.0 